requirements.txt
aioredis
python-dotenv
click
aiodine==1.*
pydantic